+++ lib/src/http/managers/channel_manager.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:http/http.dart' show MultipartFile;
 import 'package:nyxx/src/builders/builder.dart';
+import 'package:nyxx/src/builders/channel/group_dm.dart';
 import 'package:nyxx/src/builders/channel/stage_instance.dart';
 import 'package:nyxx/src/builders/channel/thread.dart';
 import 'package:nyxx/src/builders/invite.dart';
@@ -607,7 +608,7 @@ class ChannelManager extends ReadOnlyManager<Channel> {
 
       request = MultipartRequest(
         route,
-        method: 'PATCH',
+        method: 'POST',
         jsonPayload: jsonEncode(payload),
         files: files,
       );
@@ -824,5 +825,23 @@ class ChannelManager extends ReadOnlyManager<Channel> {
     await client.httpHandler.executeSafe(request);
 
     stageInstanceCache.remove(channelId);
+  }
+
+  Future<void> addRecipient(Snowflake channelId, Snowflake userId, DmRecipientBuilder builder) async {
+    final route = HttpRoute()
+      ..channels(id: channelId.toString())
+      ..recipients(id: userId.toString());
+    final request = BasicRequest(route, method: 'PUT', body: jsonEncode(builder.build()));
+
+    await client.httpHandler.executeSafe(request);
+  }
+
+  Future<void> removeRecipient(Snowflake channelId, Snowflake userId) async {
+    final route = HttpRoute()
+      ..channels(id: channelId.toString())
+      ..recipients(id: userId.toString());
+    final request = BasicRequest(route, method: 'DELETE');
+
+    await client.httpHandler.executeSafe(request);
   }
 }