+++ lib/src/http/route.dart
@@ -67,7 +67,7 @@ class HttpRoutePart {
 
 /// A parameter in a [HttpRoutePart].
 ///
-/// {@template http_route_part}
+/// {@template http_route_param}
 /// This is not a query parameter, it is a parameter encoded in the path of the request itself, such
 /// as the id of a guild in `/guilds/0123456789`.
 /// {@endtemplate}
@@ -311,4 +311,7 @@ extension RouteHelpers on HttpRoute {
 
   /// Adds the [`avatar-decorations`](https://discord.com/developers/docs/reference#image-formatting-cdn-endpoints) part to this [HttpRoute].
   void avatarDecorations({String? id}) => add(HttpRoutePart('avatar-decorations', [if (id != null) HttpRouteParam(id)]));
+
+  /// Adds the [`recipients`](https://discord.com/developers/docs/resources/channel#group-dm-add-recipient) part to this [HttpRoute].
+  void recipients({String? id}) => add(HttpRoutePart('recipients', [if (id != null) HttpRouteParam(id)]));
 }